ASoC: imx-hdmi-dma: Fix channel mapping ALSA <-> CEA, performance tweak
The speaker allocation defined by CEA-861 is different from what ALSA.
This patch changes the order of the samples to honour this. The four
channel mode is disabled, because this would require to run the hardware
in six channel mode in fill in silence on the unused channels. Further-
more the patch optimizes the inner copy loop for speed.