Revert "MLK-10199 ARM: clk-imx6q: set enet pll rate to 125Mhz"
This reverts commit d24a5df30aed6e85e19e9bfe92f184b933debbdb.

This was a global change that waas added specifically for a single
board the sabreauto cpu2 board.  I had worked around it by patching
our dts file to re-clock the enet pll to the reference required by
our phy.  For some cards this re-clock happened at the same time
as the pcie initialization and could cause erratic behaviour by
some cards causing them to hang or not be detected.

I am reverting this commit as our kernel is not meant to support
this board anyways, and this patch is not accepted in the mainline
kernel.

Conflicts:
	arch/arm/mach-imx/clk-imx6q.c